re organized project directories
