+++ android/src/main/AndroidManifest.xml
@@ -5,49 +5,49 @@
     <application>
 
         <activity
-            android:name=".SmileIDSmartSelfieEnrollmentActivity"
+            android:name=".products.selfie.SmileIDSmartSelfieEnrollmentActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
             tools:ignore="DiscouragedApi" />
 
         <activity
-            android:name=".SmileIDSmartSelfieAuthenticationActivity"
+            android:name=".products.selfie.SmileIDSmartSelfieAuthenticationActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
             tools:ignore="DiscouragedApi" />
 
         <activity
-            android:name=".enhanced.SmileIDSmartSelfieEnrollmentEnhancedActivity"
+            android:name=".products.enhancedselfie.SmileIDSmartSelfieEnrollmentEnhancedActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
             tools:ignore="DiscouragedApi" />
 
         <activity
-            android:name=".enhanced.SmileIDSmartSelfieAuthenticationEnhancedActivity"
+            android:name=".products.enhancedselfie.SmileIDSmartSelfieAuthenticationEnhancedActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
             tools:ignore="DiscouragedApi" />
 
         <activity
-            android:name=".SmileIDDocumentVerificationActivity"
+            android:name=".products.document.SmileIDDocumentVerificationActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
             tools:ignore="DiscouragedApi" />
 
         <activity
-            android:name=".SmileIDBiometricKYCActivity"
+            android:name=".products.biometric.SmileIDBiometricKYCActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
             tools:ignore="DiscouragedApi" />
 
         <activity
-            android:name=".SmileIDEnhancedDocumentVerificationActivity"
+            android:name=".products.enhanceddocv.SmileIDEnhancedDocumentVerificationActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustResize"
