+++ android/src/main/kotlin/com/smileidentity/flutter/products/enhancedselfie/SmileIDSmartSelfieAuthenticationEnhanced.kt
@@ -1,11 +1,14 @@
-package com.smileidentity.flutter
+package com.smileidentity.flutter.products.enhancedselfie
 
 import SmartSelfieCaptureResult
 import SmileIDProductsResultApi
 import android.content.Context
 import androidx.compose.runtime.Composable
 import com.smileidentity.SmileID
-import com.smileidentity.compose.SmartSelfieAuthentication
+import com.smileidentity.compose.SmartSelfieAuthenticationEnhanced
+import com.smileidentity.flutter.mapper.pathList
+import com.smileidentity.flutter.mapper.toMap
+import com.smileidentity.flutter.views.SmileComposablePlatformView
 import com.smileidentity.results.SmileIDResult
 import com.smileidentity.util.randomUserId
 import io.flutter.plugin.common.BinaryMessenger
@@ -14,48 +17,48 @@ import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
 import kotlinx.collections.immutable.toImmutableMap
 
-internal class SmileIDSmartSelfieAuthentication private constructor(
+internal class SmileIDSmartSelfieAuthenticationEnhanced private constructor(
     context: Context,
     viewId: Int,
     messenger: BinaryMessenger,
     args: Map<String, Any?>,
     private val api: SmileIDProductsResultApi,
-) : SmileSelfieComposablePlatformView(context, VIEW_TYPE_ID, viewId, messenger, args) {
+) : SmileComposablePlatformView(context, VIEW_TYPE_ID, viewId, messenger, args) {
     companion object {
-        const val VIEW_TYPE_ID = "SmileIDSmartSelfieAuthentication"
+        const val VIEW_TYPE_ID = "SmileIDSmartSelfieAuthenticationEnhanced"
     }
 
     @Composable
     override fun Content(args: Map<String, Any?>) {
         val extraPartnerParams = args["extraPartnerParams"] as? Map<String, String> ?: emptyMap()
-        SmileID.SmartSelfieAuthentication(
+        SmileID.SmartSelfieAuthenticationEnhanced(
             userId = args["userId"] as? String ?: randomUserId(),
             allowNewEnroll = args["allowNewEnroll"] as? Boolean ?: false,
-            allowAgentMode = args["allowAgentMode"] as? Boolean ?: false,
             showAttribution = args["showAttribution"] as? Boolean ?: true,
             showInstructions = args["showInstructions"] as? Boolean ?: true,
             skipApiSubmission = args["skipApiSubmission"] as? Boolean ?: false,
             extraPartnerParams = extraPartnerParams.toImmutableMap(),
         ) {
             when (it) {
-                is SmileIDResult.Error -> api.onSmartSelfieAuthenticationResult(
-                    successResultArg = null,
-                    errorResultArg = it.throwable.message
-                        ?: "Unknown error with Smart Selfie Authentication",
-                ) {}
-
                 is SmileIDResult.Success -> {
-                    val result = SmartSelfieCaptureResult(
-                        selfieFile = it.data.selfieFile.absolutePath,
-                        livenessFiles = it.data.livenessFiles.pathList(),
-                        apiResponse = it.data.apiResponse?.toMap(),
-                    )
+                    val result =
+                        SmartSelfieCaptureResult(
+                            selfieFile = it.data.selfieFile.absolutePath,
+                            livenessFiles = it.data.livenessFiles.pathList(),
+                            apiResponse = it.data.apiResponse?.toMap(),
+                        )
 
-                    api.onSmartSelfieAuthenticationResult(
+                    api.onSmartSelfieAuthenticationEnhancedResult(
                         successResultArg = result,
                         errorResultArg = null,
                     ) {}
                 }
+
+                is SmileIDResult.Error -> api.onSmartSelfieAuthenticationEnhancedResult(
+                    successResultArg = null,
+                    errorResultArg = it.throwable.message
+                        ?: "Unknown error with Smart Selfie Authentication Enhanced",
+                ) {}
             }
         }
     }
@@ -66,7 +69,7 @@ internal class SmileIDSmartSelfieAuthentication private constructor(
     ) : PlatformViewFactory(StandardMessageCodec.INSTANCE) {
         override fun create(context: Context, viewId: Int, args: Any?): PlatformView {
             @Suppress("UNCHECKED_CAST")
-            return SmileIDSmartSelfieAuthentication(
+            return SmileIDSmartSelfieAuthenticationEnhanced(
                 context,
                 viewId,
                 messenger,
